+++ commands/core/upgrade.go
@@ -18,18 +18,18 @@ package core
 import (
 	"context"
 
-	"github.com/arduino/arduino-cli/arduino/cores"
-
 	"github.com/arduino/arduino-cli/arduino"
+	"github.com/arduino/arduino-cli/arduino/cores"
 	"github.com/arduino/arduino-cli/arduino/cores/packagemanager"
 	"github.com/arduino/arduino-cli/commands"
+	"github.com/arduino/arduino-cli/commands/internal/instances"
 	rpc "github.com/arduino/arduino-cli/rpc/cc/arduino/cli/commands/v1"
 )
 
 // PlatformUpgrade FIXMEDOC
 func PlatformUpgrade(ctx context.Context, req *rpc.PlatformUpgradeRequest, downloadCB rpc.DownloadProgressCB, taskCB rpc.TaskProgressCB) (*rpc.PlatformUpgradeResponse, error) {
 	upgrade := func() (*cores.PlatformRelease, error) {
-		pme, release := commands.GetPackageManagerExplorer(req.GetInstance())
+		pme, release := instances.GetPackageManagerExplorer(req.GetInstance())
 		if pme == nil {
 			return nil, &arduino.InvalidInstanceError{}
 		}
