+++ commands/instances.go
@@ -21,7 +21,6 @@ import (
 	"net/url"
 	"path/filepath"
 	"strings"
-	"sync"
 
 	"github.com/arduino/arduino-cli/arduino"
 	"github.com/arduino/arduino-cli/arduino/cores"
@@ -34,10 +33,10 @@ import (
 	"github.com/arduino/arduino-cli/arduino/resources"
 	"github.com/arduino/arduino-cli/arduino/sketch"
 	"github.com/arduino/arduino-cli/arduino/utils"
+	"github.com/arduino/arduino-cli/commands/internal/instances"
 	"github.com/arduino/arduino-cli/configuration"
 	"github.com/arduino/arduino-cli/i18n"
 	rpc "github.com/arduino/arduino-cli/rpc/cc/arduino/cli/commands/v1"
-	"github.com/arduino/arduino-cli/version"
 	paths "github.com/arduino/go-paths-helper"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
@@ -45,90 +44,6 @@ import (
 )
 
 var tr = i18n.Tr
-
-// CoreInstance is an instance of the Arduino Core Services. The user can
-// instantiate as many as needed by providing a different configuration
-// for each one.
-type CoreInstance struct {
-	pm *packagemanager.PackageManager
-	lm *librariesmanager.LibrariesManager
-}
-
-// coreInstancesContainer has methods to add an remove instances atomically.
-type coreInstancesContainer struct {
-	instances      map[int32]*CoreInstance
-	instancesCount int32
-	instancesMux   sync.Mutex
-}
-
-// instances contains all the running Arduino Core Services instances
-var instances = &coreInstancesContainer{
-	instances:      map[int32]*CoreInstance{},
-	instancesCount: 1,
-}
-
-// GetInstance returns a CoreInstance for the given ID, or nil if ID
-// doesn't exist
-func (c *coreInstancesContainer) GetInstance(id int32) *CoreInstance {
-	c.instancesMux.Lock()
-	defer c.instancesMux.Unlock()
-	return c.instances[id]
-}
-
-// AddAndAssignID saves the CoreInstance and assigns a unique ID to
-// retrieve it later
-func (c *coreInstancesContainer) AddAndAssignID(i *CoreInstance) int32 {
-	c.instancesMux.Lock()
-	defer c.instancesMux.Unlock()
-	id := c.instancesCount
-	c.instances[id] = i
-	c.instancesCount++
-	return id
-}
-
-// RemoveID removes the CoreInstance referenced by id. Returns true
-// if the operation is successful, or false if the CoreInstance does
-// not exist
-func (c *coreInstancesContainer) RemoveID(id int32) bool {
-	c.instancesMux.Lock()
-	defer c.instancesMux.Unlock()
-	if _, ok := c.instances[id]; !ok {
-		return false
-	}
-	delete(c.instances, id)
-	return true
-}
-
-// GetPackageManager returns a PackageManager. If the package manager is not found
-// (because the instance is invalid or has been destroyed), nil is returned.
-// Deprecated: use GetPackageManagerExplorer instead.
-func GetPackageManager(instance *rpc.Instance) *packagemanager.PackageManager {
-	i := instances.GetInstance(instance.GetId())
-	if i == nil {
-		return nil
-	}
-	return i.pm
-}
-
-// GetPackageManagerExplorer returns a new package manager Explorer. The
-// explorer holds a read lock on the underlying PackageManager and it should
-// be released by calling the returned "release" function.
-func GetPackageManagerExplorer(req *rpc.Instance) (explorer *packagemanager.Explorer, release func()) {
-	pm := GetPackageManager(req)
-	if pm == nil {
-		return nil, nil
-	}
-	return pm.NewExplorer()
-}
-
-// GetLibraryManager returns the library manager for the given instance.
-func GetLibraryManager(req *rpc.Instance) *librariesmanager.LibrariesManager {
-	i := instances.GetInstance(req.GetId())
-	if i == nil {
-		return nil
-	}
-	return i.lm
-}
 
 func installTool(pm *packagemanager.PackageManager, tool *cores.ToolRelease, downloadCB rpc.DownloadProgressCB, taskCB rpc.TaskProgressCB) error {
 	pme, release := pm.NewExplorer()
@@ -146,50 +61,11 @@ func installTool(pm *packagemanager.PackageManager, tool *cores.ToolRelease, dow
 
 // Create a new CoreInstance ready to be initialized, supporting directories are also created.
 func Create(req *rpc.CreateRequest, extraUserAgent ...string) (*rpc.CreateResponse, error) {
-	instance := &CoreInstance{}
-
-	// Setup downloads directory
-	downloadsDir := configuration.DownloadsDir(configuration.Settings)
-	if downloadsDir.NotExist() {
-		err := downloadsDir.MkdirAll()
-		if err != nil {
-			return nil, &arduino.PermissionDeniedError{Message: tr("Failed to create downloads directory"), Cause: err}
-		}
+	inst, err := instances.Create(extraUserAgent...)
+	if err != nil {
+		return nil, err
 	}
-
-	// Setup data directory
-	dataDir := configuration.DataDir(configuration.Settings)
-	packagesDir := configuration.PackagesDir(configuration.Settings)
-	if packagesDir.NotExist() {
-		err := packagesDir.MkdirAll()
-		if err != nil {
-			return nil, &arduino.PermissionDeniedError{Message: tr("Failed to create data directory"), Cause: err}
-		}
-	}
-
-	// Create package manager
-	userAgent := "arduino-cli/" + version.VersionInfo.VersionString
-	for _, ua := range extraUserAgent {
-		userAgent += " " + ua
-	}
-	instance.pm = packagemanager.NewBuilder(
-		dataDir,
-		configuration.PackagesDir(configuration.Settings),
-		downloadsDir,
-		dataDir.Join("tmp"),
-		userAgent,
-	).Build()
-	instance.lm = librariesmanager.NewLibraryManager(
-		dataDir,
-		downloadsDir,
-	)
-
-	// Save instance
-	instanceID := instances.AddAndAssignID(instance)
-
-	return &rpc.CreateResponse{
-		Instance: &rpc.Instance{Id: instanceID},
-	}, nil
+	return &rpc.CreateResponse{Instance: inst}, nil
 }
 
 // Init loads installed libraries and Platforms in CoreInstance with specified ID,
@@ -205,8 +81,8 @@ func Init(req *rpc.InitRequest, responseCallback func(r *rpc.InitResponse)) erro
 	if reqInst == nil {
 		return &arduino.InvalidInstanceError{}
 	}
-	instance := instances.GetInstance(reqInst.GetId())
-	if instance == nil {
+	instance := req.GetInstance()
+	if !instances.IsValid(instance) {
 		return &arduino.InvalidInstanceError{}
 	}
 
@@ -295,7 +171,7 @@ func Init(req *rpc.InitRequest, responseCallback func(r *rpc.InitResponse)) erro
 		// after reinitializing an instance after installing or uninstalling a core.
 		// If this is not done the information of the uninstall core is kept in memory,
 		// even if it should not.
-		pmb, commitPackageManager := instance.pm.NewBuilder()
+		pmb, commitPackageManager := instances.GetPackageManager(instance).NewBuilder()
 
 		// Load packages index
 		for _, URL := range allPackageIndexUrls {
@@ -390,7 +266,7 @@ func Init(req *rpc.InitRequest, responseCallback func(r *rpc.InitResponse)) erro
 		commitPackageManager()
 	}
 
-	pme, release := instance.pm.NewExplorer()
+	pme, release := instances.GetPackageManagerExplorer(instance)
 	defer release()
 
 	for _, err := range pme.LoadDiscoveries() {
@@ -403,7 +279,7 @@ func Init(req *rpc.InitRequest, responseCallback func(r *rpc.InitResponse)) erro
 		pme.IndexDir,
 		pme.DownloadDir,
 	)
-	instance.lm = lm
+	_ = instances.SetLibraryManager(instance, lm) // should never fail
 
 	// Load libraries
 	for _, pack := range pme.GetPackages() {
@@ -485,7 +361,7 @@ func Init(req *rpc.InitRequest, responseCallback func(r *rpc.InitResponse)) erro
 
 // Destroy FIXMEDOC
 func Destroy(ctx context.Context, req *rpc.DestroyRequest) (*rpc.DestroyResponse, error) {
-	if ok := instances.RemoveID(req.GetInstance().GetId()); !ok {
+	if ok := instances.Delete(req.GetInstance()); !ok {
 		return nil, &arduino.InvalidInstanceError{}
 	}
 	return &rpc.DestroyResponse{}, nil
@@ -494,7 +370,7 @@ func Destroy(ctx context.Context, req *rpc.DestroyRequest) (*rpc.DestroyResponse
 // UpdateLibrariesIndex updates the library_index.json
 func UpdateLibrariesIndex(ctx context.Context, req *rpc.UpdateLibrariesIndexRequest, downloadCB rpc.DownloadProgressCB) error {
 	logrus.Info("Updating libraries index")
-	lm := GetLibraryManager(req.GetInstance())
+	lm := instances.GetLibraryManager(req.GetInstance())
 	if lm == nil {
 		return &arduino.InvalidInstanceError{}
 	}
@@ -523,7 +399,7 @@ func UpdateLibrariesIndex(ctx context.Context, req *rpc.UpdateLibrariesIndexRequ
 
 // UpdateIndex FIXMEDOC
 func UpdateIndex(ctx context.Context, req *rpc.UpdateIndexRequest, downloadCB rpc.DownloadProgressCB) error {
-	if instances.GetInstance(req.GetInstance().GetId()) == nil {
+	if !instances.IsValid(req.GetInstance()) {
 		return &arduino.InvalidInstanceError{}
 	}
